SITE-UPDATE
This was a major change. I removed the webpages as previously mentioned in my last css update earlier today. These are the majority of the changes and will continue to stick with 2 pages and will set up a blog style update page on my daily progress and anything else to further go back on.